backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Greg Baranov <gbaranov99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
psycopg2 = "^2.8"
django-toolbelt = "^0.0.1"
djangorestframework = "^3.10"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
